Work to get app up and running, move App.js to root and fix references, disgnate default page via AppRegistry
