+++ App.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import App from './App';
-import ScreenSelection from './ScreenSelection';
-import { StackNavigator } from 'react-navigation';
+import ScreenSelection from './src/ScreenSelection';
+import { StackNavigator } from './react-navigation';
 
 import renderer from 'react-test-renderer';
