+++ app.json
@@ -2,6 +2,6 @@
   "expo": {
     "sdkVersion": "21.0.0",
     "name": "HighFive",
-    "icon": "highfive/handprint.png"
+    "icon": "./handprint.png"
   }
 }
