+++ src/EndScreen.js
@@ -3,7 +3,6 @@ import { StyleSheet, Text, View} from 'react-native';
 import { StackNavigator } from 'react-navigation';
 import Container from './Container';
 import Button from './Button';
-import App from './App';
 
 
 export default class EndScreen extends Component {
@@ -21,6 +20,4 @@ export default class EndScreen extends Component {
 
 };
 }
-  const styles = StyleSheet.create({
-
-});
+const styles = StyleSheet.create({});
