+++ src/ScreenSelection.js
@@ -3,7 +3,6 @@ import { StyleSheet, Text, View} from 'react-native';
 import { StackNavigator } from 'react-navigation';
 import Container from './Container';
 import Button from './Button';
-import App from './App';
 
 
 export default class ScreenSelection extends Component {
